ref: Handle version push in the registry to remove circular dependencies
